+++ src/main/java/racingcar/control/input/StandardInput.java
@@ -5,12 +5,12 @@ import racingcar.view.View;
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
-public abstract class InputControl {
+public abstract class StandardInput {
     private static final Scanner scanner = new Scanner(System.in);
 
     abstract protected View view();
 
-    public int receiveInt() {
+    public int getInt() {
         this.view().render();
         try {
             return scanner.nextInt();
